% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_retrieve.R
\name{cds_retrieve}
\alias{cds_retrieve}
\title{cds_retrieve}
\usage{
cds_retrieve(dataset, request, target = "", request_id = NA)
}
\arguments{
\item{dataset}{Dataset to extract data from.}

\item{request}{JSON string containing request parameters.}

\item{target}{Filename to save data to.}

\item{request_id}{Previous request ID returned by function.}
}
\value{
List containing request ID and name of data file downloaded, NA returned if timeout
}
\description{
Function to retrieve data from the Copernicus Climate Change Service (C3S) Climate Data Store
}
\note{
API Key read from ~/.cdsapirc
}
\examples{
dataset <- "reanalysis-era5-single-levels"
request <- "{
            'variable': 'mean_sea_level_pressure',
            'grid': ['0.1', '0.1'], 'product_type': 'reanalysis',
            'year': '2010', 'month': '08', 'day': '15', 'time': '12:00',
            'area': '45.0/-15.0/70.0/25.0',
            'format': 'netcdf'
            }"
result <- cds_retrieve( dataset, request)
# If timeout the following can be used to try again without generating new request
result <- cds_retrieve( dataset, request, request_id = result$request_id)

}
